fix(hostd): drop useless i64::from on c_long syscall nr (clippy 1.95 on Linux)
clippy::useless_conversion: syscall_name_to_nr returns libc::c_long which
is i64 on 64-bit Linux, so i64::from(nr) is a no-op there. The jailer's
seccomp.rs only runs on 64-bit Linux microVMs; insert nr directly. Pre-
existing latent lint (identical in the former mvm-jailer-lite) surfaced by
CI's clippy 1.95.0 on the Linux target — invisible to a macOS dev clippy
since the file is cfg(target_os=linux).

+++ crates/mvm-hostd/src/jailer/seccomp.rs
@@ -155,7 +155,7 @@ pub fn apply(spec: &ConfinementSpec) -> Result<(), JailerError> {
                 "unknown syscall name {name:?}; extend BRIDGE_SYSCALLS in mvm-jailer-lite::seccomp"
             ))
         })?;
-        rules.insert(i64::from(nr), vec![]);
+        rules.insert(nr, vec![]);
     }
     // SeccompAction::Trap (vs Errno(EACCES)) is intentional: the
     // bridge sidecar is expected to be killed by SIGSYS on a forbidden
